+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/GeneralIT.java
@@ -59,7 +59,6 @@ import io.swagger.client.api.WorkflowsApi;
 import io.swagger.client.model.DockstoreTool;
 import io.swagger.client.model.DockstoreTool.TopicSelectionEnum;
 import io.swagger.client.model.Entry;
-import io.swagger.client.model.Image;
 import io.swagger.client.model.PublishRequest;
 import io.swagger.client.model.SourceFile;
 import io.swagger.client.model.Tag;
@@ -333,24 +332,6 @@ public class GeneralIT extends BaseIT {
         return tool;
     }
 
-    private DockstoreTool createManualDockerHubTool() {
-        DockstoreTool tool = new DockstoreTool();
-        tool.setMode(DockstoreTool.ModeEnum.MANUAL_IMAGE_PATH);
-        tool.setName("dockstore-whalesay-2");
-        tool.setNamespace("dockstoretestuser");
-        tool.setRegistryString(Registry.DOCKER_HUB.getDockerPath());
-        tool.setDefaultDockerfilePath("/Dockerfile");
-        tool.setDefaultCwlPath("/Dockstore.cwl");
-        tool.setDefaultWdlPath("/Dockstore.wdl");
-        tool.setDefaultCWLTestParameterFile("/test.cwl.json");
-        tool.setDefaultWDLTestParameterFile("/test.wdl.json");
-        tool.setIsPublished(false);
-        // This actually exists: https://bitbucket.org/DockstoreTestUser/dockstore-whalesay-2/src/master/
-        tool.setGitUrl("git@bitbucket.org:DockstoreTestUser/dockstore-whalesay-2.git");
-        tool.setToolname("alternate");
-        tool.setPrivateAccess(false);
-        return tool;
-    }
 
     DockstoreTool createManualGitLabTool() {
         final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
@@ -366,73 +347,6 @@ public class GeneralIT extends BaseIT {
         return tool;
     }
 
-    private DockstoreTool registerManualGitHubContainerRegistryToolAndAddTag() {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        ContainersApi toolApi = new ContainersApi(webClient);
-        ContainertagsApi toolTagsApi = new ContainertagsApi(webClient);
-        DockstoreTool tool = new DockstoreTool();
-        tool.setMode(DockstoreTool.ModeEnum.MANUAL_IMAGE_PATH);
-
-        // This image is used for the tool: https://ghcr.io/homebrew/core/python/3.9:3.9.6
-        tool.setRegistryString(Registry.GITHUB_CONTAINER_REGISTRY.getDockerPath());
-        tool.setNamespace("homebrew");
-        tool.setName("core/python/3.9");
-        tool.setDefaultDockerfilePath("/Dockerfile");
-        tool.setDefaultCwlPath("/Dockstore.cwl");
-        tool.setDefaultWdlPath("/Dockstore.wdl");
-        tool.setDefaultCWLTestParameterFile("/test.cwl.json");
-        tool.setDefaultWDLTestParameterFile("/test.wdl.json");
-        tool.setIsPublished(false);
-        // This actually exists: https://bitbucket.org/DockstoreTestUser/dockstore-whalesay-2/src/master/
-        tool.setGitUrl("git@bitbucket.org:DockstoreTestUser/dockstore-whalesay-2.git");
-        tool.setPrivateAccess(false);
-
-        tool = toolApi.registerManual(tool);
-
-        // Add the 3.9.6 tag for the ghcr.io/homebrew/core/python/3.9 tool
-        List<Tag> tags = new ArrayList<>();
-        Tag tag = new Tag();
-        tag.setName("3.9.6");
-        tag.setReference("master");
-        tags.add(tag);
-        toolTagsApi.addTags(tool.getId(), tags);
-        tool = toolApi.refresh(tool.getId());
-        return tool;
-    }
-
-    private DockstoreTool registerManualAmazonECRToolAndAddTag() {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        ContainersApi toolApi = new ContainersApi(webClient);
-        ContainertagsApi toolTagsApi = new ContainertagsApi(webClient);
-        DockstoreTool tool = new DockstoreTool();
-        tool.setMode(DockstoreTool.ModeEnum.MANUAL_IMAGE_PATH);
-
-        // This image is used for the tool: public.ecr.aws/ubuntu/ubuntu:18.04
-        tool.setRegistryString(Registry.AMAZON_ECR.getDockerPath());
-        tool.setNamespace("ubuntu");
-        tool.setName("ubuntu");
-        tool.setDefaultDockerfilePath("/Dockerfile");
-        tool.setDefaultCwlPath("/Dockstore.cwl");
-        tool.setDefaultWdlPath("/Dockstore.wdl");
-        tool.setDefaultCWLTestParameterFile("/test.cwl.json");
-        tool.setDefaultWDLTestParameterFile("/test.wdl.json");
-        tool.setIsPublished(false);
-        // This actually exists: https://bitbucket.org/DockstoreTestUser/dockstore-whalesay-2/src/master/
-        tool.setGitUrl("git@bitbucket.org:DockstoreTestUser/dockstore-whalesay-2.git");
-        tool.setPrivateAccess(false);
-
-        tool = toolApi.registerManual(tool);
-
-        // Add the 18.04 tag for the public.ecr.aws/ubuntu/ubuntu tool
-        List<Tag> tags = new ArrayList<>();
-        Tag tag = new Tag();
-        tag.setName("18.04");
-        tag.setReference("master");
-        tags.add(tag);
-        toolTagsApi.addTags(tool.getId(), tags);
-        tool = toolApi.refresh(tool.getId());
-        return tool;
-    }
 
     /**
      * Tests adding tags to a manually registered container
@@ -857,17 +771,6 @@ public class GeneralIT extends BaseIT {
         return tool;
     }
 
-    private DockstoreTool addDockerHubTag(DockstoreTool tool, ContainertagsApi toolTagsApi, ContainersApi toolApi) {
-        List<Tag> tags = new ArrayList<>();
-        Tag tag = new Tag();
-        tag.setName("latest");
-        tag.setReference("master");
-        tags.add(tag);
-        toolTagsApi.addTags(tool.getId(), tags);
-        tool = toolApi.refresh(tool.getId());
-        return tool;
-    }
-
     private DockstoreTool addGitLabTag(DockstoreTool tool, ContainertagsApi toolTagsApi, ContainersApi toolApi) {
         List<Tag> tags = new ArrayList<>();
         Tag tag = new Tag();
@@ -1107,29 +1010,6 @@ public class GeneralIT extends BaseIT {
         assertEquals(count, count2);
     }
 
-    @Test
-    public void testGrabChecksumFromDockerHub() {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        ContainersApi toolApi = new ContainersApi(webClient);
-        ContainertagsApi toolTagsApi = new ContainertagsApi(webClient);
-        DockstoreTool tool = createManualDockerHubTool();
-
-        tool.setDefaultDockerfilePath("/Dockerfile");
-        tool.setDefaultCwlPath("/Docstore.cwl");
-        tool = toolApi.registerManual(tool);
-
-        tool = addDockerHubTag(tool, toolTagsApi, toolApi);
-        List<Tag> tags = toolApi.getContainer(tool.getId(), null).getWorkflowVersions();
-        verifyChecksumsAreSaved(tags);
-
-        // Check for case where user deletes tag and creates new one of same name.
-        // Check that the new imageid and checksums are grabbed on refresh. Also check the old images have been deleted.
-        refreshAfterDeletedTag(toolApi, tool, tags);
-        testingPostgres.runUpdateStatement("update tool set name = 'thisnamedoesnotexist' where giturl = 'git@bitbucket.org:dockstoretestuser2/dockstore-whalesay-2.git'");
-        toolApi.refresh(tool.getId());
-        List<Tag> updatedTags = toolApi.getContainer(tool.getId(), null).getWorkflowVersions();
-        verifyChecksumsAreSaved(updatedTags);
-    }
 
     @Test
     public void testAnnotatedGitHubTag() {
@@ -1198,18 +1078,6 @@ public class GeneralIT extends BaseIT {
         assertFalse("Parsing should not have failed", failed);
     }
 
-    private void refreshAfterDeletedTag(ContainersApi toolApi, DockstoreTool tool, List<Tag> tags) {
-        String imageID = tags.get(0).getImages().get(0).getImageID();
-
-        final long count = testingPostgres.runSelectStatement("select count(*) from image", long.class);
-        testingPostgres.runUpdateStatement("update image set image_id = 'dummyid'");
-        assertEquals("dummyid", toolApi.getContainer(tool.getId(), null).getWorkflowVersions().get(0).getImages().get(0).getImageID());
-        toolApi.refresh(tool.getId());
-        final long count2 = testingPostgres.runSelectStatement("select count(*) from image", long.class);
-        assertEquals(imageID, toolApi.getContainer(tool.getId(), null).getWorkflowVersions().get(0).getImages().get(0).getImageID());
-        assertEquals(count, count2);
-    }
-
     @Test
     public void testGrabChecksumFromGitLab() {
         final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
@@ -1234,57 +1102,6 @@ public class GeneralIT extends BaseIT {
         verifyChecksumsAreSaved(updatedTags);
     }
 
-    @Test
-    public void testGrabChecksumFromGitHubContainerRegistry() {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        ContainersApi toolApi = new ContainersApi(webClient);
-        DockstoreTool tool = registerManualGitHubContainerRegistryToolAndAddTag();
-        List<Tag> tags = toolApi.getContainer(tool.getId(), null).getWorkflowVersions();
-        verifyChecksumsAreSaved(tags);
-
-        // Check for case where user deletes tag and creates new one of same name.
-        // Check that the new imageid and checksums are grabbed on refresh. Also check the old images have been deleted.
-        refreshAfterDeletedTag(toolApi, tool, tags);
-
-        // mimic getting a registry being slow/not responding and verify we do not delete the image information we already have by going to an invalid url.
-        testingPostgres.runUpdateStatement("update tool set name = 'thisnamedoesnotexist' where id=" + tool.getId());
-        toolApi.refresh(tool.getId());
-        List<Tag> updatedTags = toolApi.getContainer(tool.getId(), null).getWorkflowVersions();
-        verifyChecksumsAreSaved(updatedTags);
-    }
-
-    @Test
-    public void testGrabChecksumFromAmazonECR() {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        ContainersApi toolApi = new ContainersApi(webClient);
-        DockstoreTool tool = registerManualAmazonECRToolAndAddTag();
-        List<Tag> tags = toolApi.getContainer(tool.getId(), null).getWorkflowVersions();
-        verifyChecksumsAreSaved(tags);
-
-        // Check for case where user deletes tag and creates new one of same name.
-        // Check that the new imageid and checksums are grabbed on refresh. Also check the old images have been deleted.
-        refreshAfterDeletedTag(toolApi, tool, tags);
-
-        // mimic getting a registry being slow/not responding and verify we do not delete the image information we already have by going to an invalid url.
-        testingPostgres.runUpdateStatement("update tool set name = 'thisnamedoesnotexist' where id=" + tool.getId());
-        toolApi.refresh(tool.getId());
-        List<Tag> updatedTags = toolApi.getContainer(tool.getId(), null).getWorkflowVersions();
-        verifyChecksumsAreSaved(updatedTags);
-    }
-
-    private void verifyChecksumsAreSaved(List<Tag> tags) {
-        assertTrue("The list of tags should not be empty", tags.size() > 0);
-        for (Tag tag : tags) {
-            List<Image> images = tag.getImages();
-            assertTrue("Tag should have an image", images.size() > 0);
-            String hashType = tag.getImages().get(0).getChecksums().get(0).getType();
-            String checksum = tag.getImages().get(0).getChecksums().get(0).getChecksum();
-            assertNotNull(hashType);
-            assertFalse(hashType.isEmpty());
-            assertNotNull(checksum);
-            assertFalse(checksum.isEmpty());
-        }
-    }
 
     /**
      * Tests that if a tool has a tag with mismatching tag name and tag reference, and it is set as the default tag
